+++ CheeseQuest2/game/menu/LoadMenu.java
@@ -12,7 +12,7 @@ public class LoadMenu extends Menu {
 
     public static final String LOAD_GAME = "Load game";
     // public static final String NEW_GAME = "New game";
-    public static final String[] RETURN_TO_PREVIOUS_MENU = {"i","r"};
+    public static final String[] RETURN_TO_PREVIOUS_MENU = {"return","r"};
     public static final String[] CREATE_OPTIONS = {"create","c","create new game","create game","create new"};
     public static final String[] LOAD_OPTIONS = {"load","l"};
     public static final String[] DELETE_OPTIONS = {"delete","d"};