+++ CheeseQuest2/game/system/TextUtility.java
@@ -1,5 +1,7 @@
 package game.system;
 
+import java.util.*;
+
 /**
  * Has text utility functions
  */
@@ -74,5 +76,33 @@ public abstract class TextUtility {
      */
     public String splitCamelCaseToString(String string) {
         return String.join(" ", splitCamelCase(string));
+    }
+
+    /**
+     * [sortByLongest description]
+     * https://stackoverflow.com/questions/36292300/how-to-sort-an-arraylist-by-length-of-strings-in-the-array
+     * @param ArrayList<String> inList [description]
+     */
+    public void sortByLongest(ArrayList<String> inList) {
+        ArrayList<String> outList = new ArrayList<String>();
+        while (!inList.isEmpty()) {
+            String longest = "";
+            for (String string : inList) {
+                if (string.length() > longest.length()) {
+                    longest = string;
+                }
+            }
+            outList.add(longest);
+            while (inList.contains(longest)) {
+                inList.remove(longest);
+            }
+        }
+        inList.addAll(outList);
+    }
+
+    public String[] sortByLongest(String[] inArray) {
+        ArrayList<String> outList =new ArrayList<String>(Arrays.asList(inArray));
+        sortByLongest(outList);
+        return outList.toArray(new String[0]);
     }
 }