+++ CheeseQuest2/CHANGELOG.md
@@ -4,6 +4,16 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 
 
+[v0.8.0] - September 10, 2017
+----------------------------
+### Added
+- sortByLongest(String[]) and sortByLongest(ArrayList<String>) for TextUtility
+### Fixed
+- Order of elements in arrayList for inputStartsWithChoice() does not matter.
+    - Uses sortByLongest(ArrayList<String>)
+
+
+
 [v0.7.0] - September 9, 2017
 ---------------------------
 ### Added